code cleanup and translate

+++ idler.go
@@ -2,7 +2,9 @@ package idler
 
 type Idle struct{}
 
-func (f *Idle) GetIdleTime() float64 {
+// GetIdleTime returns the number of seconds of inactivity of user activity
+// IDLE time in seconds
+func (f *Idle) GetIdleTime() int {
 	if f == nil {
 		return 0
 	}
